requirements.txt
requests
beautifulsoup4
pillow
streamlit
httpx